Phidata/requirements.txt
phidata
duckduckgo_search
python-dotenv
yfinance